Updated to improve compatibility between Cocoa and Cocoa Touch.

+++ es_ios_utils/universal/ESCollectionCategories.h
@@ -59,7 +59,7 @@ typedef void(^ESNSManagedObjectBlock)(NSManagedObject*);
   @property(readonly) NSSet*          asSet;
   @property(readonly) NSMutableSet*   asMutableSet;
   @property(readonly) NSMutableArray* asMutableArray;
-  -(int)firstIndexWhereKeyPath:(id)kp isEqual:(id)o;
+  -(NSInteger)firstIndexWhereKeyPath:(id)kp isEqual:(id)o;
   -(int)countObjectsWhereKeyPath:(id)kp isEqual:(id)o;
 @end
